Form basics almost done
Have made the basic layout for the website. Banner at top now stretches across screen. Added a nav to go back (doesn't go anywhere) Added the fields for the forms, along with the forms. Most of the left side fields are text, whereas the right side ones are selects or buttons.